+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Version 2.0 preview work in progress
 - new #33: Added widget Focuser (VladimirShestakov)
 - new #34: Added widget PagePreview with style (VladimirShestakov)
 - new #35: Added widget PageNavigation in the basic package and layout boolive for widget the part (VladimirShestakov)
+- new #43: Added handler forms in the package "basic" and in /Interfaces (VladimirShestakov)
 - new #44: Added indication that the object exists (VladimirShestakov)
 - new: Added object the Option for automatic views (VladimirShestakov)
 - new: Added top menu and sidebar menu in the boolive layout (VladimirShestakov)